removed retrieverconfig from modules.config and added individual parameters to filters constructor
